+++ src/Database/PostgreSQL/PQTypes/Model/ForeignKey.hs
@@ -25,7 +25,7 @@ data ForeignKey = ForeignKey {
 , fkOnDelete   :: ForeignKeyAction
 , fkDeferrable :: Bool
 , fkDeferred   :: Bool
-, fkValidated  :: Bool -- ^ Set to 'True' if foreign key is created as NOT VALID
+, fkValidated  :: Bool -- ^ Set to 'False' if foreign key is created as NOT VALID
                        -- and not validated afterwards.
 } deriving (Eq, Ord, Show)